refactor(server): dedupe TextDecoder stream consumption
Extract the repeated `new TextDecoder()` + `ReadableStream` chunk-loop
into shared helpers in `utils/text-stream.ts`:

- `readStreamAsText(stream)` for the consume-to-single-string pattern
  (used in `app-page-cache.ts` ISR write and `pages-page-response.ts`
  ISR cache write).
- `readStreamAsTextWithLimit(stream, maxBytes, onLimitExceeded)` for
  the size-limited pattern (used in `app-server-action-execution.ts`
  and `pages-node-compat.ts`). Each caller passes its own throwing
  callback so error types (`Error` vs `PagesBodyParseError`) and
  status codes are preserved.

Sites with additional load-bearing logic intentionally left inline:
- `rsc-stream-hints.ts` (line-buffered TransformStream with carry)
- `app-ssr-stream.ts` (raw-byte accumulator + flight hint fix; HTML
  flush transform with `fixPreloadAs`)
- `shims/fetch-cache.ts` (size-limited cache-key body canonicalisation
  with mixed string/Uint8Array stream and tee handling)
